src/requirements.txt
python_igraph==0.11.3
PyYAML==6.0.1
tabulate==0.9.0
